:zap: Refactor fetchAPIData
