--- src/util/fetchAPIData.ts
@@ -1,1 +1,0 @@
-export * from '@/util/fetch/server-util';
